move TContainerImage ad GetFullContainerImageName.ts to dmt common
