+++ web/packages/plugins/shared/common/src/index.tsx
@@ -19,4 +19,5 @@ export * from './domain/Tree'
 export * from './components/TreeView'
 export * from './components'
 export * from './utils/truncatePathString'
+export * from './utils/GetFullContainerImageName'
 export { ErrorBoundary } from './utils/ErrorBoundary'
