+++ web/packages/plugins/shared/common/src/utils/GetFullContainerImageName.ts
@@ -1,4 +1,4 @@
-import { TContainerImage } from '../Types'
+import { TContainerImage } from '../types'
 
 export const getFullContainerImageName = (image: TContainerImage): string => {
   return `${image.registryName}/${image.imageName}:${image.version}`
